requirements.txt
aiohttp==3.6.2
async-lru==1.0.2
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
idna==2.9
kuriyama==0.3.3
Logbook==1.5.3
multidict==4.7.5
pydantic==1.5.1
PyYAML==5.3.1
yarl==1.4.2
